Only reintroduce the explicit casts from ee35d4250bfdbad5ca53424d

+++ src/os.c
@@ -294,7 +294,7 @@ static char *read_from_shell_command(char *result, size_t sizeof_result, const c
     if ((fp = popen(command, "r")) == NULL) {
         return NULL;
     }
-    if (fgets(result, sizeof_result, fp) == NULL) {
+    if (fgets(result, (int) sizeof_result, fp) == NULL) {
         pclose(fp);
         fprintf(stderr, "Command [%s] failed]\n", command);
         return NULL;
